feat: hello.atlas now work
