+++ src/atlas_runtime/visitor.rs
@@ -1,46 +1,59 @@
-use crate::atlas_frontend::parser::ast::{
-    BinaryExpression, DoExpression, Expression, FieldAccessExpression, FunctionCall,
-    FunctionExpression, IdentifierNode, IfElseNode, IndexExpression, MatchExpression,
-    NewObjectExpression, UnaryExpression, VariableDeclaration,
-};
-use crate::atlas_memory::vm_data::VMData;
+use crate::{atlas_frontend::parser::ast::AstLiteral, atlas_memory::vm_data::VMData};
 
-pub type Program = Vec<Expression>;
+use super::{
+    errors::RuntimeResult, AstBinaryOpExpr, AstBlock, AstCallExpr, AstExpr, AstFieldAccessExpr,
+    AstFunction, AstIdentifier, AstIfElseExpr, AstIndexingExpr, AstLetExpr, AstMatchExpr,
+    AstProgram, AstUnaryOpExpr,
+};
 
 //TODO: visit() should return a Result<VMData, crate::errors::RuntimeError>
 pub trait Visitor<'visitor> {
     type CallBack;
     // Entry point
-    fn visit(&mut self, program: &'visitor Program) -> VMData;
+    fn visit(&mut self, program: &'visitor AstProgram, entry_point: &str) -> RuntimeResult<VMData>;
 
     // Expressions
-    fn visit_expression(&mut self, expression: &'visitor Expression) -> VMData;
-    fn visit_binary_expression(&mut self, expression: &'visitor BinaryExpression) -> VMData;
-    fn visit_unary_expression(&mut self, expression: &'visitor UnaryExpression) -> VMData;
+    fn visit_expression(&mut self, expression: &'visitor AstExpr) -> RuntimeResult<VMData>;
+    fn visit_binary_expression(
+        &mut self,
+        expression: &'visitor AstBinaryOpExpr,
+    ) -> RuntimeResult<VMData>;
+    fn visit_unary_expression(
+        &mut self,
+        expression: &'visitor AstUnaryOpExpr,
+    ) -> RuntimeResult<VMData>;
     fn visit_function_expression(
         &mut self,
-        function_expression: &'visitor FunctionExpression,
-    ) -> VMData;
-    fn visit_function_call(&mut self, function_call: &'visitor FunctionCall) -> VMData;
-    fn visit_index_expression(&mut self, index_expression: &'visitor IndexExpression) -> VMData;
+        function_expression: &'visitor AstFunction,
+    ) -> RuntimeResult<VMData>;
+    fn visit_function_call(
+        &mut self,
+        function_call: &'visitor AstCallExpr,
+    ) -> RuntimeResult<VMData>;
+    fn visit_index_expression(
+        &mut self,
+        index_expression: &'visitor AstIndexingExpr,
+    ) -> RuntimeResult<VMData>;
     fn visit_field_access_expression(
         &mut self,
-        field_access_expression: &'visitor FieldAccessExpression,
-    ) -> VMData;
-    fn visit_new_object_expression(
-        &mut self,
-        new_object_expression: &'visitor NewObjectExpression,
-    ) -> VMData;
-
+        field_access_expression: &'visitor AstFieldAccessExpr,
+    ) -> RuntimeResult<VMData>;
     // Variables and Identifiers
     fn visit_variable_declaration(
         &mut self,
-        variable_declaration: &'visitor VariableDeclaration,
-    ) -> VMData;
-    fn visit_identifier(&mut self, identifier: &'visitor IdentifierNode) -> VMData;
+        variable_declaration: &'visitor AstLetExpr,
+    ) -> RuntimeResult<VMData>;
+    fn visit_identifier(&mut self, identifier: &'visitor AstIdentifier) -> RuntimeResult<VMData>;
+    fn visit_literal(&mut self, literal: &'visitor AstLiteral) -> RuntimeResult<VMData>;
 
     // Control flow
-    fn visit_if_else_node(&mut self, if_else_node: &'visitor IfElseNode) -> VMData;
-    fn visit_do_expression(&mut self, do_expression: &'visitor DoExpression) -> VMData;
-    fn visit_match_expression(&mut self, match_expression: &'visitor MatchExpression) -> VMData;
+    fn visit_if_else_node(
+        &mut self,
+        if_else_node: &'visitor AstIfElseExpr,
+    ) -> RuntimeResult<VMData>;
+    fn visit_block_expression(&mut self, block: &'visitor AstBlock) -> RuntimeResult<VMData>;
+    fn visit_match_expression(
+        &mut self,
+        match_expression: &'visitor AstMatchExpr,
+    ) -> RuntimeResult<VMData>;
 }
